# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VendorIQ API
  version: 0.1.0
  summary: Vendor management & market intelligence for construction tendering.
  description: |
    The public REST contract of VendorIQ. **Every UI action goes through this API** — the
    frontend holds no business logic (brief §2), so anything a screen can do is an operation
    below, and any other product can do the same with an API key.

    This document is hand-written and is the source of truth: `apps/api` serves it verbatim at
    `/api/openapi.json`, `/api/openapi.yaml` and `/api/docs`, and route handlers are validated
    against it rather than generating it. Changing it is a contract change — see
    `CONTRIBUTING.md`.

    ### Conventions

    * **Ids** are UUIDs, stable across systems. `external_ref` on vendors and projects maps to
      a foreign system's own id.
    * **Errors** always use one envelope: `{"error": {"code", "message", "details"}}`.
    * **Collections** always return `{"items", "total", "page", "page_size"}`.
    * **Money** is AZN. The Rev4 methodology sheet labels turnover USD; the data is AZN and the
      system stores AZN (spec §16).
    * **Language** — user-facing text carries `_az` and `_en` fields; the server never returns a
      single pre-translated sentence for the UI, it returns keys or both languages.
    * **Field codes** (`A.1` … `G.7`, tables `C.t1`, `C.t2`, `G.t1`) come from spec Appendix A;
...
      model version.
...
  license:
    name: Proprietary
  contact:
...
servers:
...
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /applications/{application_id}/evaluation:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    get:
      tags: [applications]
      operationId: getEvaluation
      summary: The evaluation sheet
      description: Every criterion with its raw indicator, evidence document code, rubric cell and points.
      responses:
        "200":
          description: Evaluation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Evaluation" }
    put:
      tags: [applications]
      operationId: putEvaluation
      summary: Save the officer's rubric scores
      description: Rubric cells are integers 0–3 keyed by criterion code. Saving recomputes and stores `computed`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RubricInput" }
      responses:
        "200":
          description: Saved and recomputed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Evaluation" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: The application's model version is locked against edits at this status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /applications/{application_id}/compute:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    post:
      tags: [applications]
      operationId: computeScore
      summary: Score without saving
      description: |
        Powers the live evaluation screen: the client sends the rubric cells it currently has on
        screen and gets per-criterion points, group totals, total, KO and class back. **Nothing is
        persisted** — this is what keeps scoring out of the frontend.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ComputeRequest" }
      responses:
        "200":
          description: Computed scores.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScoreResult" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /applications/{application_id}/decide:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    post:
      tags: [applications]
      operationId: decideApplication
      summary: Approve, reject or request information
      description: |
        `approve` is refused below the pass mark or on a knock-out failure. A justification is
        mandatory for `reject` and `request_info`. Emits `vendor.prequalified` on approval.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionInput" }
      responses:
        "200":
          description: Decision recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDetail" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: The decision is not allowed from this state or score.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /applications/{application_id}/second-evaluator:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    put:
      tags: [applications]
      operationId: putSecondEvaluation
      summary: Record the second evaluator's rubric set
      description: Optional mode (spec §10.3). The response flags criteria where the two sets differ by more than one point.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RubricInput" }
      responses:
        "200":
          description: Saved, with the divergence list.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecondEvaluation" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /cycles/{cycle_id}/export-summary.xlsx:
    parameters: [{ $ref: "#/components/parameters/CycleId" }]
    get:
      tags: [applications]
      operationId: exportCommissionSummaryXlsx
      summary: Commission summary for signature (Excel)
      description: Reproduces the layout of the workbook sheet "5. Nəticə Xülasəsi".
      parameters: [{ $ref: "#/components/parameters/LocaleParam" }]
      responses:
        "200": { $ref: "#/components/responses/XlsxFile" }
  /cycles/{cycle_id}/export-summary.pdf:
    parameters: [{ $ref: "#/components/parameters/CycleId" }]
    get:
      tags: [applications]
      operationId: exportCommissionSummaryPdf
      summary: Commission summary for signature (PDF)
      parameters: [{ $ref: "#/components/parameters/LocaleParam" }]
      responses:
        "200": { $ref: "#/components/responses/PdfFile" }

  # ────────────────────────────────── cycles ─────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: vendoriq_session
      description: httpOnly session cookie issued by the auth endpoints. Mutating requests also require the `X-CSRF-Token` header.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Machine credential with scopes; see `/integrations/api-keys`.
  parameters:
...
    LocaleParam:
      name: locale
      in: query
      description: Language of the generated file's headings.
      schema: { type: string, enum: [az, en], default: az }
...
    ApplicationId:
      name: application_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CycleId:
      name: cycle_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Forbidden:
      description: The role or scope does not allow this operation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationError:
      description: The payload failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    XlsxFile:
      description: An Excel workbook.
      content:
        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
          schema: { type: string, format: binary }
    PdfFile:
      description: A PDF document.
      content:
        application/pdf:
          schema: { type: string, format: binary }
  schemas:
...
    ErrorEnvelope:
      type: object
      description: The one error shape every endpoint returns.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: Stable machine-readable code.
              enum:
                - bad_request
                - unauthenticated
                - forbidden
                - not_found
                - conflict
                - payload_too_large
                - unsupported_media_type
                - validation_error
                - rate_limited
                - internal_error
            message:
              type: string
              description: English text for logs and developers, never shown verbatim to a vendor.
            details:
              type: object
              additionalProperties: true
              description: Structured context — failed checks, offending field codes, retry-after.
      examples:
        - error:
            code: conflict
            message: Pre-submission check failed.
            details:
              checks:
                mandatory_fields: false
                mandatory_documents: true
                knock_out_answers: true
...
    ApplicationStatus:
      type: string
      enum:
        [
          invited,
          in_progress,
          submitted,
          under_review,
          information_requested,
          prequalified,
          rejected,
          withdrawn,
        ]
    ScoreClass:
      type: string
      enum: [A, B, C, D, F, KO]
      description: "A ≥ 90, B ≥ 80, C ≥ 70, D ≥ 60, F < 60. `KO` overrides the total."
...
    ObservationSource:
      type: string
      enum: [registry, api, document, portal, excel, manual]
      description: "Trust rank 1–5: registry 1, api 2, document 3, portal/excel 4, manual 5."
...
    DecisionKind:
      type: string
      enum: [approve, reject, request_info]
...
    CriterionKind:
      type: string
      enum: [rubric, bands, thresh, ongoing, leadtime]
...
    Application:
      type: object
      required: [id, vendor_id, cycle_id, status]
      properties:
        id: { type: string, format: uuid }
        vendor_id: { type: string, format: uuid }
        vendor_name: { type: string }
        cycle_id: { type: string, format: uuid }
        cycle_name: { type: string }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        submitted_at: { type: [string, "null"], format: date-time }
        total: { type: [number, "null"] }
        cls: { $ref: "#/components/schemas/ScoreClass" }
        decision: { $ref: "#/components/schemas/DecisionKind" }
        decided_at: { type: [string, "null"], format: date-time }
        evaluator_name: { type: [string, "null"] }
        is_demo: { type: boolean }
...
    ApplicationDetail:
      allOf:
        - $ref: "#/components/schemas/Application"
        - type: object
          properties:
            scoring_model_version: { type: string }
            answers:
              type: object
              additionalProperties: true
              description: Answers by field code; tables as arrays.
            raw_snapshot:
              type: [object, "null"]
              additionalProperties: { type: [number, "null"] }
              description: Frozen at submission; null before it.
            rubric_scores:
              type: [object, "null"]
              additionalProperties: { type: integer, minimum: 0, maximum: 3 }
            computed: { $ref: "#/components/schemas/ScoreResult" }
            declaration: { $ref: "#/components/schemas/Declaration" }
            justification: { type: [string, "null"] }
            checks: { $ref: "#/components/schemas/SubmissionChecks" }
            score_released:
              type: boolean
              description: Vendors see the breakdown only after the commission decision.
...
    SubmissionChecks:
      type: object
      description: The pre-submission checklist shown on the declaration screen.
      required: [mandatory_fields, mandatory_documents, knock_out_answers]
      properties:
        mandatory_fields: { type: boolean }
        mandatory_documents: { type: boolean }
        knock_out_answers: { type: boolean }
        missing_field_codes:
          type: array
          items: { type: string }
        missing_document_codes:
          type: array
          items: { type: string }
    Declaration:
      type: object
      properties:
        signatory_name: { type: [string, "null"] }
        signatory_position: { type: [string, "null"] }
        agreed: { type: boolean }
        signed_at: { type: [string, "null"], format: date-time }
        stamp_file_key: { type: [string, "null"] }
...
    ScoreResult:
      type: object
      description: Engine output — the same shape the reference JS returns.
      required: [per, groups, total, ko, cls]
      properties:
        per:
          type: object
          additionalProperties: { type: number }
          description: Points per criterion code, rounded to one decimal.
        groups:
          type: object
          additionalProperties: { type: number }
          description: Points per group letter, rounded to one decimal after each addition.
        total: { type: number, description: "Sum of the group totals, rounded to one decimal." }
        ko: { type: boolean, description: False when any knock-out criterion has a raw value of 0. }
        cls: { $ref: "#/components/schemas/ScoreClass" }
        pass_mark: { type: number }
        model_version: { type: string }
    Evaluation:
      type: object
      required: [application_id, model_version, rows, computed]
      properties:
        application_id: { type: string, format: uuid }
        model_version: { type: string }
        rows:
          type: array
          items: { $ref: "#/components/schemas/EvaluationRow" }
        computed: { $ref: "#/components/schemas/ScoreResult" }
        can_approve:
          type: boolean
          description: False below the pass mark or on a knock-out failure.
        evaluator_name: { type: [string, "null"] }
    EvaluationRow:
      type: object
      required: [code, group, kind, max]
      properties:
        code: { type: string }
        group: { type: string }
        name_az: { type: string }
        name_en: { type: string }
        kind: { $ref: "#/components/schemas/CriterionKind" }
        max: { type: number }
        ko: { type: boolean }
        unit: { type: [string, "null"] }
        evidence_doc: { type: [string, "null"], description: Checklist code the officer verifies against. }
        raw_value: { type: [number, "null"] }
        raw_source: { $ref: "#/components/schemas/ObservationSource" }
        rubric_score: { type: [integer, "null"], minimum: 0, maximum: 3 }
        points: { type: number }
...
    RubricInput:
      type: object
      required: [rubric_scores]
      properties:
        rubric_scores:
          type: object
          additionalProperties: { type: integer, minimum: 0, maximum: 3 }
        evidence:
          type: object
          additionalProperties: { type: string }
          description: Optional per-criterion note about the document that was checked.
    ComputeRequest:
      type: object
      description: Score with these rubric cells without persisting anything.
      properties:
        rubric_scores:
          type: object
          additionalProperties: { type: integer, minimum: 0, maximum: 3 }
        raw_overrides:
          type: object
          additionalProperties: { type: [number, "null"] }
          description: What-if values for numeric criteria; defaults to the application's snapshot.
        model_version:
          type: string
          description: Defaults to the cycle's version.
    SecondEvaluation:
      type: object
      required: [computed, divergences]
      properties:
        computed: { $ref: "#/components/schemas/ScoreResult" }
        divergences:
          type: array
          description: Criteria where the two evaluators differ by more than one point.
          items:
            type: object
            properties:
              code: { type: string }
              first: { type: integer }
              second: { type: integer }
    DecisionInput:
      type: object
      required: [decision]
      properties:
        decision: { $ref: "#/components/schemas/DecisionKind" }
        justification:
          type: string
          description: Mandatory for `reject` and `request_info`.
        valid_months:
          type: integer
          description: Overrides the model's 12-month validity for this approval.

    # ── cycles ───────────────────────────────────────────────────────────────
...